Move quiz timer to QuizMenu component
Moved quiz timer to QuizMenu component for automatic timer start. Also created separate stylesheet for quiz timer.
