+++ src/components/styles/TimerStyles.js
@@ -1,0 +1,18 @@
+import styled from 'styled-components';
+
+const TimerStyles = styled.div`
+  .statusBar-timerButtons {
+    position: absolute;
+    transition: all 0.5s;
+    position: absolute;
+    left: 1em;
+    top: 1em;
+    z-index: 2;
+  }
+
+  .ui.inverted.red.button {
+    margin-right: 1em;
+  }
+`;
+
+export default TimerStyles;
